requirements.txt
Django==1.5.4
dj-database-url==0.2.2
dj-static==0.0.5
gunicorn==18.0
psycopg2==2.5.1
static==0.4
django-bootstrap3==1.0.0
UniversalClient==0.5.2
google-api-python-client==1.2
Pillow==2.2.1
python-dateutil==2.1
